% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{emd_palette}
\alias{emd_palette}
\title{A contrastive palette ranging from light to dark}
\usage{
emd_palette(var)
}
\arguments{
\item{var}{A categorical variable with no more than eight unique values; if
it is a factor then the color palette will follow the order of the factor levels
from light to dark}
}
\value{
A named vector of colours; if there are two colours
then they are orange (light) and blue (dark)
}
\description{
A contrastive palette ranging from light to dark
}

