% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{emd_theme}
\alias{emd_theme}
\title{The theme I use for papers}
\usage{
emd_theme(text_size = 18)
}
\arguments{
\item{text_size}{Text size (default: 18)}
}
\value{
A ggplot theme object
}
\description{
The theme I use for papers
}

