% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{factor_grouped}
\alias{factor_grouped}
\title{Assign factors to groups}
\usage{
factor_grouped(x, groups = NULL)
}
\arguments{
\item{x}{A vector (if \code{x} is not a factor vector then it
will be cast to one with the default level order)}

\item{groups}{A vector of groups of length \code{k}, where
\code{k} is the floor of
the number of levels of \code{x} divided by four, plus one if the number of levels
is not evenly divisible by four; if \code{NULL}, \code{1:k} will be used}
}
\value{
A vector of length equal to the number of levels of \code{x},
with the levels of \code{x} as labels, containing the appropriate groups for
each level: the first element of \code{groups} will be repeated four times
(or less if there are less than four levels of \code{x}),
then the second level four times (or less if there are less than eight levels of
\code{x}), and so on
}
\description{
The \code{\link{emd_colours}} used by
\code{\link{emd_palette}} only yield four distinct grey levels
or colour blind colours, and eight distinct RGB colours,
which repeat down the factor levels, in order. It's
therefore useful (usually crucial) to have a second factor that groups the first
four levels of \code{x} together, the second four, the third
four, and so on, for some factor vector \code{x}.
}

