% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{hist_area_line}
\alias{hist_area_line}
\title{Returns a ggplot containing overlapping histograms for different groups, filled
in with areas and with a thick line over top}
\usage{
hist_area_line(x, group = NULL, colour_palette = emd_palette(group),
  var_measure_name = "x", var_group_name = "group", line_width = 3,
  highlight_all = F, y_type = "count")
}
\arguments{
\item{x}{A variable to plot}

\item{group}{grouping variable}

\item{colour_palette}{A vector containing the colours for each group; if this vector is named
then the order of the names will be used to order the groups on the plot}

\item{var_measure_name}{An display label for the dependent variable}

\item{var_group_name}{An display label for the group variable}

\item{line_width}{Line width}

\item{highlight_all}{If TRUE, all groups get a thick line, otherise we leave off the top one}

\item{y_type}{"density", "ndensity", "count", or "ncount": density estimate, density estimate
scaled to maximum of one, count, or count normalized to maximum of one}
}
\value{
A ggplot plot object
}
\description{
Returns a ggplot containing overlapping histograms for different groups, filled
in with areas and with a thick line over top
}

